requirements.txt
groq>=0.4.0
python-dotenv>=1.0.0
livekit-agents>=0.9.0
livekit-plugins-groq>=0.1.0
livekit-plugins-elevenlabs>=0.7.9

# Optional: Payment processing (uncomment to enable monetization features with real Stripe integration)
# stripe>=7.0.0
